+++ packages/guideline-themes/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @frontify/guideline-themes
 
+## 1.1.0
+
+### Minor Changes
+
+-   [#1568](https://github.com/Frontify/brand-sdk/pull/1568) [`1aa19a4`](https://github.com/Frontify/brand-sdk/commit/1aa19a4c0fabddae79a1e882206a187ab466714e) Thanks [@Kenny806](https://github.com/Kenny806)! - Introduce new settings structure in sidebar-settings to allow more flexibility for settings. Old structure still supported.
+
+### Patch Changes
+
+-   [#1566](https://github.com/Frontify/brand-sdk/pull/1566) [`3889cea`](https://github.com/Frontify/brand-sdk/commit/3889cea0894955a9acdaa7315cd466eb414a119f) Thanks [@ragi96](https://github.com/ragi96)! - chore(Deps): bumping some devDependencies
+
+-   Updated dependencies [[`3889cea`](https://github.com/Frontify/brand-sdk/commit/3889cea0894955a9acdaa7315cd466eb414a119f), [`1aa19a4`](https://github.com/Frontify/brand-sdk/commit/1aa19a4c0fabddae79a1e882206a187ab466714e)]:
+    -   @frontify/app-bridge-theme@1.0.8
+    -   @frontify/sidebar-settings@0.11.0
+
 ## 1.0.4
 
 ### Patch Changes